+++ testFramework/com/twitter/intellij/pants/testFramework/OSSPantsIntegrationTest.java
@@ -64,31 +64,6 @@ abstract public class OSSPantsIntegrationTest extends PantsIntegrationTestCase {
     return stringList.stream().anyMatch(s -> s.contains(subString));
   }
 
-  protected void assertAndRunPantsMake(JUnitConfiguration runConfiguration) {
-
-    RunManager runManager = RunManager.getInstance(myProject);
-    assertTrue(runManager instanceof RunManagerImpl);
-    RunManagerImpl runManagerImpl = (RunManagerImpl) runManager;
-
-    RunnerAndConfigurationSettings runnerAndConfigurationSettings =
-      runManagerImpl.createConfiguration(runConfiguration, JUnitConfigurationType.getInstance().getConfigurationFactories()[0]);
-    runManagerImpl.addConfiguration(runnerAndConfigurationSettings, false);
-
-    // Make sure PantsMake is the one and only task before JUnit run.
-    List<BeforeRunTask> beforeRunTaskList = runManagerImpl.getBeforeRunTasks(runConfiguration);
-    assertEquals(1, beforeRunTaskList.size());
-    BeforeRunTask task = beforeRunTaskList.iterator().next();
-    assertEquals(PantsMakeBeforeRun.ID, task.getProviderId());
-
-    /*
-     * Manually invoke BeforeRunTask as {@link ExecutionManager#compileAndRun} launches another task asynchronously,
-     * and there is no way to catch that.
-     */
-    BeforeRunTaskProvider provider = BeforeRunTaskProvider.getProvider(myProject, task.getProviderId());
-    assertNotNull(String.format("Cannot find BeforeRunTaskProvider for id='%s'", task.getProviderId()), provider);
-    assertTrue(provider.executeTask(null, runConfiguration, null, task));
-  }
-
   protected void assertCompileAll() {
     PantsMakeBeforeRun runner = new PantsMakeBeforeRun(myProject);
     assertTrue(runner.executeTask(myProject));
